installation scripts at platform targets only

+++ macos/install.sh
@@ -1,20 +1,41 @@
 #!/bin/bash
 # Simple bash script to install dotfiles for vim/bash/etc.
-# This script is macOS specific.
-#
-# Usage:
-# $ ./install.sh
 #
 # author: @dencold
 #
 
-# first, figure out what directory we are in. This script is called from the main 
+# Warning message & prompt for users
+echo
+echo "************************ WARNING - DOTFILES INSTALL ************************ "
+echo
+echo "   Running this script will replace your dotfiles (.vimrc/.bashrc/etc.)   "
+echo "   and any associated config directories (such as .vim) "
+echo "   Back up previous versions if they are important to you. "
+echo ""
+echo "************************* YOU HAVE BEEN WARNED! :) ************************** "
+echo
+read -p "Are you SURE? (y/N) ==> " SURE
+
+if [[ "$SURE" != "Y" && "$SURE" != "y" ]];
+then
+    echo "Aborting dotfiles install."
+    exit 0
+fi
+
+#-----------------------------------------------------------------------------------
+# Directory Initialization
+# - sets up our required directories
+#-----------------------------------------------------------------------------------
+echo "- Creating initial directories"
+mkdir -p $HOME/src $HOME/bin $HOME/tmp
+
+# figure out what directory we are in. This script is called from the main 
 # install.sh in the root directory (dotfiles). This SO article is *very* helpful:
 # http://stackoverflow.com/questions/59895/can-a-bash-script-tell-what-directory-its-stored-in
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 echo "- Installing profiles"
-for file in .aliases .bashrc .bash_funcs .bash_profile .bash_prompt .inputrc; do
+for file in .aliases .bashrc .bash_funcs .bash_profile .bash_prompt .inputrc .tmux.conf; do
     cp -p $DIR/$file $HOME/$file;
 done;
 
@@ -37,3 +58,26 @@ done;
 # replace the placeholders in .gitconfig with user input
 sed -i -e "s/GIT_NAME/$full_name/g" $HOME/.gitconfig
 sed -i -e "s/GIT_EMAIL/$email/g" $HOME/.gitconfig
+
+#-----------------------------------------------------------------------------------
+# vim
+# - installs vim dotfiles & associated plugins
+#-----------------------------------------------------------------------------------
+echo "- Installing vim directory/plugins"
+rm -rf $HOME/.vim
+
+# first we initialize the plugin submodules, see the README for this repository
+# for more background on what is happening here.
+git submodule init ../vim/bundle
+git submodule update --init --recursive ../vim/bundle
+
+# now install our vim directory
+cp -rp ../vim $HOME/.vim
+
+# create our the directories where backups will go
+mkdir -p $HOME/.vim/backup $HOME/.vim/swp
+
+echo
+echo "!! FINISHED - dotfiles install !!"
+echo "- reminder that you may need to source your new profile"
+echo
